moved from sashabaranov/go-gpt3 to sashabaranov/go-openai
This is inspired by https://github.com/turbot/steampipe-plugin-openai/issues/9
and created a working merge from  judell's outdated branch.

This is the starting point to improve upon it,

but there is a braking change:

I removed settings = '{"prompt": "my prompt"}'
as valid, since the new chatCompletion works a little different.